Cleanup and finish split of notification flow
