+++ athena/queues/direct-message-notification.js
@@ -1,5 +1,5 @@
 // @flow
-const debug = require('debug')('athena:queue:message-notification');
+const debug = require('debug')('athena:queue:direct-message-notification');
 import { fetchPayload, createPayload } from '../utils/payloads';
 import { getDistinctActors } from '../utils/actors';
 import {
@@ -11,10 +11,9 @@ import {
   storeUsersNotifications,
   markUsersNotificationsAsNew,
 } from '../models/usersNotifications';
-import { getThreadNotificationUsers } from '../models/usersThreads';
 import { getDirectMessageThreadMembers } from '../models/usersDirectMessageThreads';
 import sentencify from '../utils/sentencify';
-import bufferNotificationEmail from './buffer-message-notification-email';
+import bufferNotificationEmail from './buffer-direct-message-notification-email';
 
 const formatAndBufferNotificationEmail = (
   recipient,
@@ -58,12 +57,7 @@ const formatAndBufferNotificationEmail = (
 const processDirectMessageNotificationQueue = job => {
   const incomingMessage = job.data.message;
   const currentUserId = job.data.userId;
-
-  // Determine what the context type should be based on the message that was sent
-  const contextType =
-    incomingMessage.threadType === 'directMessageThread'
-      ? 'DIRECT_MESSAGE_THREAD'
-      : 'THREAD';
+  const contextType = 'DIRECT_MESSAGE_THREAD';
 
   debug(
     `new job: message sent by ${currentUserId} in ${contextType.toLowerCase()}#${incomingMessage.threadId}`
@@ -118,10 +112,9 @@ const processDirectMessageNotificationQueue = job => {
         notificationPromise
           // Do the .then here so we keep the loaded data in scope
           .then(notification => {
-            const getRecipients =
-              contextType === 'DIRECT_MESSAGE_THREAD'
-                ? getDirectMessageThreadMembers(notification.context.id)
-                : getThreadNotificationUsers(notification.context.id);
+            const getRecipients = getDirectMessageThreadMembers(
+              notification.context.id
+            );
 
             debug('get recipients for notification');
             return Promise.all([notification, getRecipients]);
@@ -149,26 +142,20 @@ const processDirectMessageNotificationQueue = job => {
                 formatAndBufferNotificationEmail(
                   recipient,
                   // Return direct message thread data in the same format as normal threads
-                  contextType === 'DIRECT_MESSAGE_THREAD'
-                    ? {
-                        content: {
-                          // Contruct title out of direct message thread users
-                          title: `your conversation with ${sentencify(
-                            recipients
-                              .filter(
-                                userThread =>
-                                  userThread.userId !== recipient.userId
-                              )
-                              .map(user => user.name)
-                          )}`,
-                        },
-                        path: `messages/${thread.id}`,
-                        id: thread.id,
-                      }
-                    : {
-                        ...thread,
-                        path: `thread/${thread.id}`,
-                      },
+                  {
+                    content: {
+                      // Contruct title out of direct message thread users
+                      title: `your conversation with ${sentencify(
+                        recipients
+                          .filter(
+                            userThread => userThread.userId !== recipient.userId
+                          )
+                          .map(user => user.name)
+                      )}`,
+                    },
+                    path: `messages/${thread.id}`,
+                    id: thread.id,
+                  },
                   user,
                   message,
                   notification
